cleaned up code: renamed file folders, added comments to classes & functions, other clean up
cleaned up code: renamed file folders, added comments to classes & functions, other clean up
