+++ kala-app/src/assets/landingPage/landing-pg.jpg
