+++ kala-app/src/customHistory.tsx
@@ -1,5 +1,6 @@
 import { createBrowserHistory } from "history";
 
+// Custom browser history object to be used for app routing & navigation purposes
 const history = createBrowserHistory();
 
 export default history;
