+++ kala-app/src/fundingForm/Questions/BizQ.tsx
@@ -4,9 +4,9 @@ business industry, type and date. Data is temporarily saved to redux
 */
 import { connect } from 'react-redux';
 import React from 'react';
-import { updateFilters } from '../../redux-data/actions';
-import AppState, { Filters } from '../../redux-data/types';
-import './../fundingForm.css';
+import { updateFilters } from '../../reduxData/actions';
+import AppState, { Filters } from '../../reduxData/types';
+import '../fundingForm.css';
 import kala from '../../assets/kala_orange_solid 3.svg';
 
 // material ui form styling
