+++ kala-app/src/fundingForm/WelcomeQuestion.tsx
