+++ kala-app/src/libraryPages/categoryPages/covidSupport.tsx
@@ -2,6 +2,9 @@ import React from 'react';
 import './category.css';
 import {createResource} from './createResource';
 
+/**
+Creates the COVID Support page of the library with pertinent resources
+*/
 class CovidSupport extends React.Component<any> {
   componentDidMount() {
     window.scrollTo(0, 0);
