--- kala-app/src/library_pages/libraryHistory.tsx
@@ -1,5 +1,0 @@
-import { createBrowserHistory } from "history";
-
-const libHistory = createBrowserHistory();
-
-export default libHistory;
