+++ kala-app/src/searchPages/TaExpanded.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './searchStyle.css';
 import { connect } from "react-redux";
-import { AppState, TA } from "../redux-data/types";
+import { AppState, TA } from "../reduxData/types";
 import { ArrowBackIos } from '@material-ui/icons';
 
 // properties that belong to SearchHome
