+++ kala-app/src/searchPages/searchStyle.css
